### 2021-09-30 #136 @whedon
[PRE REVIEW]: Computational models of human social behavior and neuroscience: An open educational course and Jupyter Book to advance computational training
**Submitting author:** @shawnrhoads (<a href="http://orcid.org/0000-0003-1350-9458">Shawn A. Rhoads</a>)
**Repository:** <a href="https://github.com/shawnrhoads/gu-psyc-347" target ="_blank">https://github.com/shawnrhoads/gu-psyc-347</a>
**Version:** v1.0.0
**Managing EiC:** Jordan Gorzalski

Due to the challenges of the COVID-19 pandemic, JOSE is currently operating in a "reduced service mode". You can read more about what that means in [our blog post](https://blog.joss.theoj.org/2020/05/reopening-joss).
