+++ paper.bib
@@ -15,6 +15,7 @@ year={2017}
 @article{Guest2021,
 abstract = {Psychology endeavors to develop theories of human capacities and behaviors on the basis of a variety of methodologies and dependent measures. We argue that one of the most divisive factors in psychological science is whether researchers choose to use computational modeling of theories (over and above data) during the scientific-inference process. Modeling is undervalued yet holds promise for advancing psychological science. The inherent demands of computational modeling guide us toward better science by forcing us to conceptually analyze, specify, and formalize intuitions that otherwise remain unexamined—what we dub open theory. Constraining our inference process through modeling enables us to build explanatory and predictive theories. Here, we present scientific inference in psychology as a path function in which each step shapes the next. Computational modeling can constrain these steps, thus advancing scientific inference over and above the stewardship of experimental practice (e.g., preregistration). If psychology continues to eschew computational modeling, we predict more replicability crises and persistent failure at coherent theory building. This is because without formal modeling we lack open and transparent theorizing. We also explain how to formalize, specify, and implement a computational model, emphasizing that the advantages of modeling can be achieved by anyone with benefit to all.},
 author = {Guest, Olivia and Martin, Andrea E},
+doi = {10.1177/1745691620970585},
 journal = {Perspectives on Psychological Science},
 keywords = {challenges for scientific inference,cognitive science,computational model,in psychological,open science,path function,pedagogy,philosophy of science,psychology,replicability,reproducibility,research methods,scientific inference,theoretical},
 number = {4},
