### 2021-09-30 #136 @shawnrhoads
Thank you! I have added the DOI [referenced above](https://github.com/openjournals/jose-reviews/issues/136#issuecomment-931490754) with [this PR](https://github.com/shawnrhoads/gu-psyc-347/pull/179)
