### 2021-10-19 #146 @whedon
[REVIEW]: Computational models of human social behavior and neuroscience: An open educational course and Jupyter Book to advance computational training
**Submitting author:** @shawnrhoads (<a href="http://orcid.org/0000-0003-1350-9458">Shawn A. Rhoads</a>)
**Repository:** <a href="https://github.com/shawnrhoads/gu-psyc-347" target ="_blank">https://github.com/shawnrhoads/gu-psyc-347</a>
**Version:** v1.1.0

Due to the challenges of the COVID-19 pandemic, JOSE is currently operating in a "reduced service mode".
