### 2021-10-19 #146 @emckiernan
Hi @arokem and @djmannion: Thank you for agreeing to review for JOSE! There are individual checklists for you both at the top of this issue. Please work your way through those, feel free to ask questions or post comments here, and also open issues in the submission repository as needed. I'm here to help, so please let me know if you need anything. Thanks!
