### 2021-10-20 shawnrhoads/gu-psyc-347#180 @arokem
[JOSE] How to get help
To address the review point "Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the module 2) Report issues or problems with the module 3) Seek support".

I think that sub-point 1 (contribution) is well addressed by your CONTRIBUTING.md file. I think that the others could be improved in the README, or in the content of the Jupyterbook, or both.
