### 2021-10-20 #146 @shawnrhoads
>
>
> To this question:
>
> "Version: Does the release version given match the repository release (v1.0.0)?"
>
> There are currently no releases on the GitHub repo, although there is a `v1.0.0` tag that could easily be released through the GitHub interface, if that's needed.
>

@arokem, thanks for pointing this out! I just changed the repo settings to display the release version. Does it look right, now? https://github.com/shawnrhoads/gu-psyc-347
