### 2021-11-03 #146 @arokem
Clearly, not going fast enough 😄

I am still working through this slowly. I hope to be able to work through everything in the next couple of weeks.
