### 2021-11-04 #146 @djmannion
A general comment about the documentation (which I take to mean the repository documentation, rather than the jupyter book), which affects my response to those criteria, is that it seems to be for students in the course rather than for others who might like to adapt/use the material.

For example, it looks like the repository `README.md` is copied to become the index page of the jupyter book. That means that it has things like the course logistics that are relevant for students, but not for non-students considering using the material - and that any changes to the `README.md` to address such topics (e.g., how to build the book) would be not appropriate to show to students.

Might it be better to separate the material for those different audiences? For example, have the repository documentation (`README.md`) be about the need for the module from an educator perspective, how someone can adapt and build it for their own course, etc., and have all the student-targeted material in `docs`?
