### 2021-11-05 #146 @arokem
For this point:

> Does it describe how it has been used in the classroom or other settings, and how someone might adopt it?

I think that you could add a bit in the paper about your experience . This is also a +1 to @djmannion's comment above that the document could be improved to make this more accessible to other instructors. You might consider adding a section about how to use the materials from the instructor point of view.

This point:

> Does the paper tell the "story" of how the authors came to develop it, or what their expertise is?

Also does not currently clearly come through.
