### 2021-12-03 #146 @shawnrhoads
Hi all,

Thank you so much for reviewing these materials! These are great ideas, especially separating the documentation for different audiences (i.e., instructors versus students).

We will work on incorporating this helpful feedback and new issues submitted to the repository in the coming days and keep everyone posted on progress.

Shawn & Lin
