+++ docs/index1.md
@@ -1,0 +1,79 @@
+<a target="_blank" rel="noopener noreferrer" href="https://shawnrhoads.github.io/gu-psyc-347">![Jupyter Book Badge](https://jupyterbook.org/badge.svg)</a> <a target="_blank" rel="noopener noreferrer" href="https://github.com/shawnrhoads/gu-psyc-347/actions/workflows/deploy-book.yml">![book](https://github.com/shawnrhoads/gu-psyc-347/actions/workflows/deploy-book.yml/badge.svg)</a>
+
+# Computational Models of Human Social Behavior and Neuroscience
+Last updated: September 2021</br>
+The content in this Jupyter Book is subject to change. 
+
+## Course information 📝
+**Section**: PSYC 347-01</br>
+**Prerequisites**: PSYC 002 - Research Methods and Statistics (or equivalent); No prior programming experience necessary</br>
+**Required materials**: Only a working laptop/computer is needed! Book chapters, journal articles, and software are openly available to all students (no purchase required)</br>
+**Dates**: January 28 - May 6, 2021</br>
+**Meetings**: Tuesdays and Thursdays from 11am-12:15pm</br>
+**Location**: Online</br>
+**Syllabus**: <a target="_blank" rel="noopener noreferrer" href="https://shawnrhoads.github.io/gu-psyc-347/module-00-00_Syllabus.html">Found here</a></br>
+**Schedule**: <a target="_blank" rel="noopener noreferrer" href="https://shawnrhoads.github.io/gu-psyc-347/module-00-01_Course-Schedule.html">Found here</a>
+
+## Instructor information 👨‍🏫
+**Instructor**: Shawn A Rhoads<br>
+**Pronouns**: he/him/his<br>
+**Email**: sr1209(at)georgetown(dot)edu<br>
+**Twitter**: <a target="_blank" rel="noopener noreferrer" href="https://twitter.com/ShawnRhoads56">@ShawnRhoads56</a><br>
+**Website**: <a target="_blank" rel="noopener noreferrer" href="https://shawnrhoads.github.io">shawnrhoads.github.io</a></br>
+
+## Tutorials & Exercises 🏋️
+
+### Module 01 - Scientific Python
+
+|   |  Run  |  View  |
+| - | :---: | :----: |
+| Tutorial: Jupyter Notebooks | <a target="_blank" rel="noopener noreferrer" href="https://colab.research.google.com/github/shawnrhoads/gu-psyc-347/blob/master/docs/module-01-00_Jupyter-Notebooks-Colab.ipynb">![Open In Colab](https://colab.research.google.com/assets/colab-badge.svg)</a> | <a target="_blank" rel="noopener noreferrer" href="https://shawnrhoads.github.io/gu-psyc-347/module-01-00_Jupyter-Notebooks.html">![view](https://github.com/shawnrhoads/gu-psyc-347/actions/workflows/deploy-book.yml/badge.svg)</a> |
+| Tutorial: Python Basics | <a target="_blank" rel="noopener noreferrer" href="https://colab.research.google.com/github/shawnrhoads/gu-psyc-347/blob/master/docs/module-01-01_Intro-to-Python.ipynb">![Open In Colab](https://colab.research.google.com/assets/colab-badge.svg)</a> | <a target="_blank" rel="noopener noreferrer" href="https://shawnrhoads.github.io/gu-psyc-347/module-01-01_Intro-to-Python.html">![view](https://github.com/shawnrhoads/gu-psyc-347/actions/workflows/deploy-book.yml/badge.svg)</a> |
[…]
