+++ .github/workflows/deploy-book.yml
@@ -44,7 +44,6 @@ jobs:
         conda list
         conda activate gu-psyc-347-jb
         jupyter nbextension enable --py widgetsnbextension
-        cp README.md docs/index.md
         cp CONTRIBUTING.md docs/module-00-06_Contributing.md
         jupyter-book build docs
 
