### 2021-12-07 1c5af5e
updated badges and added info for instructors
