### 2021-12-07 f51579f
updated contributing information with PR guidelines
