+++ CONTRIBUTING.md
@@ -20,11 +20,25 @@ Here are some ways you can help me accomplish this goal!
 - If I had inadvertently omitted credit for any content generated by others, please also let me know.
 - If you have an idea for a tutorial or a new module to include (especially related to the [growing list of readings](https://shawnrhoads.github.io/gu-psyc-347/module-00-03_Reading-List.html)), please either [open a new issue](https://github.com/shawnrhoads/gu-psyc-347/issues/new?title=Tutorial%20Idea:%20[INSERT%20TITLE]&body=Please%20suggest%20a%20tutorial%20idea%20here%20with%20any%20relevant%20resources.) and/or submit a pull request directly to the repository on GitHub.
 
-If you add new files, adhere to the following file naming heuristic:
+## Submitting New Content
+
+All new content should meet the goals listed above. To add new pages to the Jupyter Book, please complete the following steps:
+- [Fork the repository](https://docs.github.com/en/get-started/quickstart/fork-a-repo)
+- Create a new branch from master and begin making your edits
+- Once your edits are complete, submit a [pull request](https://docs.github.com/en/get-started/quickstart/contributing-to-projects#making-a-pull-request) to request a review of your additions/edits 
+
+The pull request should list how the new content (1) enhances the course materials and (2) meets the relevant goals. Relevant resources or readings should also be included (if applicable). 
+
+If you add new files, please adhere to the following file naming heuristic:
 - `docs/Module-**-**_TutorialName.ipynb`
 - `docs/Module-**-**_TutorialName.md`
 
-This website is hosted on [GitHub](https://github.com/shawnrhoads/gu-psyc-347) and built using [Jupyter Book](https://jupyterbook.org/intro.html), which creates a [Jekyll](https://jekyllrb.com/) website from [Markdown](https://daringfireball.net/projects/markdown/) and [Jupyter Notebooks](https://jupyter.org/).
+Please also remember to add any new pages as an entry in the table of contents `docs/_toc.yml`, so that they are built into the Jupyter Book.
+
+This website is hosted on [GitHub](https://github.com/shawnrhoads/gu-psyc-347) and built using [Jupyter Book](https://jupyterbook.org/intro.html), which creates a [Jekyll](https://jekyllrb.com/) website from [Markdown](https://daringfireball.net/projects/markdown/) and [Jupyter Notebooks](https://jupyter.org/). The continuous integration workflow for building the  Jupyter Book can be found [here](https://github.com/shawnrhoads/gu-psyc-347/blob/master/.github/workflows/deploy-book.yml). This workflow runs (1) when a new pull request is submitted to check compatibility with the master branch and (2) any time a new change is merged to the master branch. The latter run rebuilds the Jupyter Book with the updated content.
+
+Refer to [this page](https://docs.github.com/en/get-started/quickstart/contributing-to-projects) for comprehensive tips on contributing to projects on GitHub.
+
 
 ## Contributor Code of Conduct
 
