### 2021-12-07 b806fef
updates to contributing and asking for help
added issue templates for easy reporting of bugs, requests for enhancements, and question submissions

(related to #180)
