+++ .github/ISSUE_TEMPLATE/enhancement-template.yml
@@ -1,0 +1,50 @@
+name: Enhancement Report
+title: "[Enhancement]: "
+labels: ["enhancement"]
+assignees:
+  - shawnrhoads
+description:
+  Use this template to new content or features in the Jupyter Book or other course materials
+body:
+  - type: markdown
+    attributes:
+      value: Thanks for taking the time to fill out this enhancement report! Please describe the enhancement in simple words.
+  - type: input
+    id: webpage
+    attributes:
+      label: Webpage
+      description: Please paste the URL of the webpage (i.e., links to Jupyter Book, Google Colab, or GitHub repository) for which you are requesting an enhancement.
+      placeholder: ex. https://shawnrhoads.github.io/gu-psyc-347/
+    validations:
+      required: false
+  - type: textarea
+    id: description
+    attributes:
+      label: Description
+      description: What is the enhancement (e.g., descriptions, code, feature, design, etc.)? Please tell us what you see. What do you want to see?
+      placeholder: Tell us how to improve!
+      value: ""
+    validations:
+      required: true
+  - type: checkboxes
+    attributes:
+      label: Please check any relevant labels
+      options:
+        - label: content
+        - label: data viz
+        - label: documentation
+        - label: jupyter book
+        - label: notebook code
+  - type: textarea
+    attributes:
+      label: Other information
+      description:
+        Anything else that might be helpful? E.g., related issues/enhancement, suggestions how to improve, links/references with relevant examples
+  - type: checkboxes
+    id: terms
+    attributes:
+      label: Code of Conduct
+      description: By submitting this issue, you agree to follow our [Code of Conduct](https://shawnrhoads.github.io/gu-psyc-347/module-00-06_Contributing.html#contributor-code-of-conduct)
+      options:
+        - label: I agree to follow this project's Code of Conduct
+          required: true
