### 2021-12-07 775f000
updated README to reflect all new changes
changes (related to #180)
- reorganized README structure
- added info for students + instructors
- added extra info for contributing
- added info for submitting questions
