### 2021-12-07 450133b
added instructions for building jupyter book locally
