+++ .github/ISSUE_TEMPLATE/enhancement-template.yml
@@ -9,6 +9,7 @@ body:
   - type: markdown
     attributes:
       value: Thanks for taking the time to fill out this enhancement report! Please describe the enhancement in simple words.
+
   - type: input
     id: webpage
     attributes:
@@ -17,15 +18,16 @@ body:
       placeholder: ex. https://shawnrhoads.github.io/gu-psyc-347/
     validations:
       required: false
+
   - type: textarea
     id: description
     attributes:
       label: Description
       description: What is the enhancement (e.g., descriptions, code, feature, design, etc.)? Please tell us what you see. What do you want to see?
       placeholder: Tell us how to improve!
-      value: ""
     validations:
       required: true
+
   - type: checkboxes
     attributes:
       label: Please check any relevant labels
@@ -35,11 +37,13 @@ body:
         - label: documentation
         - label: jupyter book
         - label: notebook code
+
   - type: textarea
     attributes:
       label: Other information
       description:
         Anything else that might be helpful? E.g., related issues/enhancement, suggestions how to improve, links/references with relevant examples
+
   - type: checkboxes
     id: terms
     attributes:
