### 2021-12-14 450e9c2
minor updates to text and links
