+++ README.md
@@ -4,13 +4,13 @@
 
 ## For Students 👩‍🎓👨‍🎓
 
-All course information relevant for students can be found on the <a target="_blank" rel="noopener noreferrer" href="https://shawnrhoads.github.io/gu-psyc-347">course homepage</a> and <a target="_blank" rel="noopener noreferrer" href="https://shawnrhoads.github.io/gu-psyc-347/module-00-00_Syllabus.html">syllabus</a>.
+All course information relevant for students can be found on the <a target="_blank" rel="noopener noreferrer" href="https://shawnrhoads.github.io/gu-psyc-347">course homepage</a> and <a target="_blank" rel="noopener noreferrer" href="https://shawnrhoads.github.io/gu-psyc-347/module-00-00_Syllabus.html">syllabus</a>. Students are free to follow along on their own or with an instructor (see below).
 
 ## For Instructors 👩‍🏫👨‍🏫
 
 This repository contains materials for a semester-long educational course (see 14-week <a target="_blank" rel="noopener noreferrer" href="https://shawnrhoads.github.io/gu-psyc-347/module-00-01_Course-Schedule.html">schedule</a>) and <a target="_blank" rel="noopener noreferrer" href="https://shawnrhoads.github.io/gu-psyc-347">Jupyter Book</a> that provides introductory training in specifying, implementing, and interpreting computational models that characterize human social behavior and neuroscience. This course was designed and taught by <a target="_blank" rel="noopener noreferrer" href="https://shawnrhoads.github.io">Shawn A Rhoads</a> in the spring of 2021 (reach him via <a target="_blank" rel="noopener noreferrer" href="mailto:sr1209(at)georgetown(dot)edu">email</a>, <a target="_blank" rel="noopener noreferrer" href="https://github.com/shawnrhoads">GitHub</a>, or <a target="_blank" rel="noopener noreferrer" href="https://twitter.com/ShawnRhoads56">Twitter</a> with any questions).
 
-The initial release contains the following modules compiled with the Jupyter Book (and accompanying Google Colaboratory Notebooks): 
+This release contains the following modules compiled with the Jupyter Book (and accompanying Google Colaboratory Notebooks): 
 
 ### Module 01 - Scientific Python
 
@@ -44,22 +44,24 @@ The initial release contains the following modules compiled with the Jupyter Boo
 | Tutorial: Prosocial Learning | <a target="_blank" rel="noopener noreferrer" href="https://colab.research.google.com/github/shawnrhoads/gu-psyc-347/blob/master/docs/module-04-00_Social-Learning.ipynb">![Open In Colab](https://colab.research.google.com/assets/colab-badge.svg)</a> | <a target="_blank" rel="noopener noreferrer" href="https://shawnrhoads.github.io/gu-psyc-347/module-04-00_Social-Learning.html">![view](https://jupyterbook.org/badge.svg)</a> |
 | Exercise | <a target="_blank" rel="noopener noreferrer" href="https://colab.research.google.com/github/shawnrhoads/gu-psyc-347/blob/master/docs/module-04-01_Prosocial-RL-Exercises.ipynb">![Open In Colab](https://colab.research.google.com/assets/colab-badge.svg)</a> | <a target="_blank" rel="noopener noreferrer" href="https://shawnrhoads.github.io/gu-psyc-347/module-04-01_Prosocial-RL-Exercises.html">![view](https://jupyterbook.org/badge.svg)</a> |
 
[…]
