+++ _sources/index.md
@@ -1,7 +1,6 @@
-<a target="_blank" rel="noopener noreferrer" href="https://shawnrhoads.github.io/gu-psyc-347">![Jupyter Book Badge](https://jupyterbook.org/badge.svg)</a> <a target="_blank" rel="noopener noreferrer" href="https://github.com/shawnrhoads/gu-psyc-347/actions/workflows/deploy-book.yml">![book](https://github.com/shawnrhoads/gu-psyc-347/actions/workflows/deploy-book.yml/badge.svg)</a>
+<a target="_blank" rel="noopener noreferrer" href="https://shawnrhoads.github.io/gu-psyc-347">![Jupyter Book Badge](https://jupyterbook.org/badge.svg)</a> <a target="_blank" rel="noopener noreferrer" href="https://github.com/shawnrhoads/gu-psyc-347/actions/workflows/deploy-book.yml">![book](https://github.com/shawnrhoads/gu-psyc-347/actions/workflows/deploy-book.yml/badge.svg)</a> ![GitHub last update](https://img.shields.io/github/last-commit/shawnrhoads/gu-psyc-347?color=blue&label=last%20update)
 
 # Computational Models of Human Social Behavior and Neuroscience
-Last updated: September 2021</br>
 The content in this Jupyter Book is subject to change. 
 
 ## Course information 📝
@@ -21,39 +20,41 @@ The content in this Jupyter Book is subject to change.
 **Twitter**: <a target="_blank" rel="noopener noreferrer" href="https://twitter.com/ShawnRhoads56">@ShawnRhoads56</a><br>
 **Website**: <a target="_blank" rel="noopener noreferrer" href="https://shawnrhoads.github.io">shawnrhoads.github.io</a></br>
 
+Other instructors are encouraged to adopt, remix, transform, and/or build upon the following materials for educational purposes under a <a rel="license" target="_blank" rel="noopener noreferrer" href="http://creativecommons.org/licenses/by-sa/4.0/">Creative Commons Attribution-ShareAlike 4.0 International License</a>. Information for instructors can be found in the <a target="_blank" rel="noopener noreferrer" href="https://github.com/shawnrhoads/gu-psyc-347">README documentation</a>.
+
 ## Tutorials & Exercises 🏋️
 
 ### Module 01 - Scientific Python
 
 |   |  Run  |  View  |
 | - | :---: | :----: |
-| Tutorial: Jupyter Notebooks | <a target="_blank" rel="noopener noreferrer" href="https://colab.research.google.com/github/shawnrhoads/gu-psyc-347/blob/master/docs/module-01-00_Jupyter-Notebooks-Colab.ipynb">![Open In Colab](https://colab.research.google.com/assets/colab-badge.svg)</a> | <a target="_blank" rel="noopener noreferrer" href="https://shawnrhoads.github.io/gu-psyc-347/module-01-00_Jupyter-Notebooks.html">![view](https://github.com/shawnrhoads/gu-psyc-347/actions/workflows/deploy-book.yml/badge.svg)</a> |
-| Tutorial: Python Basics | <a target="_blank" rel="noopener noreferrer" href="https://colab.research.google.com/github/shawnrhoads/gu-psyc-347/blob/master/docs/module-01-01_Intro-to-Python.ipynb">![Open In Colab](https://colab.research.google.com/assets/colab-badge.svg)</a> | <a target="_blank" rel="noopener noreferrer" href="https://shawnrhoads.github.io/gu-psyc-347/module-01-01_Intro-to-Python.html">![view](https://github.com/shawnrhoads/gu-psyc-347/actions/workflows/deploy-book.yml/badge.svg)</a> |
[…]
