+++ _sources/index.md
@@ -59,7 +59,7 @@ Other instructors are encouraged to adopt, remix, transform, and/or build upon t
 ## Acknowledgments 🙏
 I am grateful to <a target="_blank" rel="noopener noreferrer" href="https://deborahstearns.blogspot.com/">Dr. Deborah Sterns</a>, <a target="_blank" rel="noopener noreferrer" href="https://www.joscelinrocha.com/">Joscelin Rocha-Hidalgo</a>, and the <a target="_blank" rel="noopener noreferrer" href="https://cndls.georgetown.edu/">Georgetown Center for New Designs in Learning & Scholarship</a> (CNDLS) for allowing me to use their teaching resources throughout the development of this course. I am also grateful to <a target="_blank" rel="noopener noreferrer" href="https://jupyter.org/">Project Jupyter</a> for making it possible to create and share these materials in a Jupyter Book. I am also grateful to <a target="_blank" rel="noopener noreferrer" href="https://github.com/gllg4009">Lin Gan</a> for her contributions and feedback in making enhancements to these materials. The format of this course was largely inspired by the <a target="_blank" rel="noopener noreferrer" href="https://mindsummerschool.org/">MIND Computational Summer School</a>, <a target="_blank" rel="noopener noreferrer" href="https://academy.neuromatch.io/">Neuromatch Academy</a>, Dr. Maximilian Risenhuber's Computational Neuroscience course, Dr. Robert C. Wilson's <a target="_blank" rel="noopener noreferrer" href="http://u.arizona.edu/~bob/web_NSCS344/">Modeling the Mind course</a>, Dr. Luke J. Chang's <a target="_blank" rel="noopener noreferrer" href="https://dartbrains.org">DartBrains Jupyter Book</a>, the research conducted in Dr. Abigail A. Marsh's <a target="_blank" rel="noopener noreferrer" href="https://aamarsh.wordpress.com/lab/">Laboratory on Social and Affective Neuroscience</a>, and the countless conversations with my colleagues about the selected topics.
 
-## Contributing 🙋‍♀️
+## Contributing 🤝
 Please visit <a target="_blank" rel="noopener noreferrer" href="https://shawnrhoads.github.io/gu-psyc-347/module-00-06_Contributing.html">this page</a> if you would like to help improve and/or expand the content in this Jupyter Book!
 
 ## Contributors ✨
