+++ .github/ISSUE_TEMPLATE/issue-template.yml
@@ -23,7 +23,7 @@ body:
     id: description
     attributes:
       label: Description
-      description: What is the issue (e.g., typo, bug, inaccurate descriptions, etc.)? Please tell us what you see. What did you expect?
+      description: What is the issue (e.g., typos, bugs, inaccurate descriptions, etc.)? Please tell us what you see. What did you expect?
       placeholder: Tell us what you see!
     validations:
       required: true
@@ -48,7 +48,7 @@ body:
   - type: dropdown
     id: browsers
     attributes:
-      label: If applicable, what browsers are you seeing the problem on?
+      label: If applicable, what browser(s) are you seeing the problem on?
       multiple: true
       options:
         - Firefox
