### 2022-01-07 9658694
added Resources page, updated info for instructors
