+++ README.md
@@ -48,11 +48,17 @@ The course structure is outlined in the <a target="_blank" rel="noopener norefer
 
 All course materials are also designed to be modular. For example, if students are knowledgeable in Python basics and Jupyter Notebook (**Module 01**), then instructors could begin with model fitting (**Module 02**). If students are knowledgeable in model fitting with maximum likelihood estimation (**Module 02**), instructors could start with reinforcement learning (**Module 03**). Instructors can adopt, remix, transform, and build upon these materials for their own course or for other educational purposes (e.g., workshops, single class sessions) under a <a rel="license" target="_blank" rel="noopener noreferrer" href="http://creativecommons.org/licenses/by-sa/4.0/">Creative Commons Attribution-ShareAlike 4.0 International License</a>.
 
-For additional background on the content covered in each tutorial (technical or theoretical), instructors are encouraged to review the references and/or resources listed at the beginning of each tutorial. In addition, instructors should familarize themselves with the <a target="_blank" rel="noopener noreferrer" href="https://shawnrhoads.github.io/gu-psyc-347/module-00-03_Reading-List.html">recommended readings</a> associated with each module and use them to introduce students to each topic, generate discussions and ideas among students, and provide relevant background for each tutorial. 
+Students have the option to interact with each tutorial notebook using one of two ways. The first option entails <a target="_blank" rel="noopener noreferrer" href="https://docs.anaconda.com/anaconda/install/">installing Anaconda 3</a> on their local computers with a <a target="_blank" rel="noopener noreferrer" href="https://raw.githubusercontent.com/shawnrhoads/gu-psyc-347/master/course-env.yml">custom environment</a> made for this course and then opening each notebook using Jupyter Notebook. Instructors can walk students through the installation steps using the instructions provided on the <a target="_blank" rel="noopener noreferrer" href="https://shawnrhoads.github.io/gu-psyc-347/module-00-04_Getting-Started.html">Getting Started</a> page. However, this can be rather difficult to implement with students' different machines, operating systems, background, etc. As a fail-safe, students also have the option to use <a target="_blank" rel="noopener noreferrer" href="https://colab.research.google.com/">Google Colaboratory</a>, which runs the code on Google's cloud servers for **free**. There is a button at the top of each tutorial page that opens each Jupyter Book page in Google Colaboratory:
+
+![Open In Colab](https://colab.research.google.com/assets/colab-badge.svg)
+
+Once students are able to run notebooks (locally or on the cloud), they can interact with all of the code in the tutorials. Instructors should guide students through all of the "tutorials", which, for example, will allow them to demonstrate to students how tweaking certain variables in real-time would change the outputs. The "exercises" at the end of each module were created for the students to apply what they have learned. They can work on these by themselves or in groups.
+
[…]
