+++ RESOURCES.md
@@ -1,0 +1,83 @@
+# Teaching Resources
+
+For additional background, instructors (and students) can review the following non-exhaustive list of resources relevant to each of the modules included in this course.
+
+### Module 01 - Scientific Python
+
+|   |  Run  |  View  |
+| - | :---: | :----: |
+| Tutorial: Jupyter Notebooks | <a target="_blank" rel="noopener noreferrer" href="https://colab.research.google.com/github/shawnrhoads/gu-psyc-347/blob/master/docs/module-01-00_Jupyter-Notebooks-Colab.ipynb">![Open In Colab](https://colab.research.google.com/assets/colab-badge.svg)</a> | <a target="_blank" rel="noopener noreferrer" href="https://shawnrhoads.github.io/gu-psyc-347/module-01-00_Jupyter-Notebooks.html">![view](https://jupyterbook.org/badge.svg)</a> |
+| Tutorial: Python Basics | <a target="_blank" rel="noopener noreferrer" href="https://colab.research.google.com/github/shawnrhoads/gu-psyc-347/blob/master/docs/module-01-01_Intro-to-Python.ipynb">![Open In Colab](https://colab.research.google.com/assets/colab-badge.svg)</a> | <a target="_blank" rel="noopener noreferrer" href="https://shawnrhoads.github.io/gu-psyc-347/module-01-01_Intro-to-Python.html">![view](https://jupyterbook.org/badge.svg)</a> |
+| Tutorial: Working with Data | <a target="_blank" rel="noopener noreferrer" href="https://colab.research.google.com/github/shawnrhoads/gu-psyc-347/blob/master/docs/module-01-02_Working-with-Data.ipynb">![Open In Colab](https://colab.research.google.com/assets/colab-badge.svg)</a> | <a target="_blank" rel="noopener noreferrer" href="https://shawnrhoads.github.io/gu-psyc-347/module-01-02_Working-with-Data.html">![view](https://jupyterbook.org/badge.svg)</a> |
+| Exercise | <a target="_blank" rel="noopener noreferrer" href="https://colab.research.google.com/github/shawnrhoads/gu-psyc-347/blob/master/docs/module-01-03_Python-Exercises.ipynb">![Open In Colab](https://colab.research.google.com/assets/colab-badge.svg)</a> | <a target="_blank" rel="noopener noreferrer" href="https://shawnrhoads.github.io/gu-psyc-347/module-01-03_Python-Exercises.html">![view](https://jupyterbook.org/badge.svg)</a> |
+
+- [Eshin Jolly - Intro to Jupyter Notebooks](https://github.com/Summer-MIND/mind_2018/tree/master/tutorials/jupyter)
+- [Python for Beginners](https://www.python.org/about/gettingstarted/)
+- [Jake Vanderplas - Data Science Handbook](https://github.com/jakevdp/PythonDataScienceHandbook)
+- [Jake VanderPlas - Whirlwind Tour of Python](https://github.com/jakevdp/WhirlwindTourOfPython) 
+- [Tal Yarkoni - Intro to Data Science w/ Python](https://github.com/tyarkoni/SSI2016)
+- [Yaroslav Halchenko - Intro to Programming for Psych/Neuro](https://github.com/dartmouth-pbs/psyc161)
+- [Collection of Jupyter Notebooks](https://github.com/jupyter/jupyter/wiki)
+- [The homogenization of scientific computing, or why Python is steadily eating other languages’ lunch](https://www.talyarkoni.org/blog/2013/11/18/the-homogenization-of-scientific-computing-or-why-python-is-steadily-eating-other-languages-lunch/)
[…]
