+++ RESOURCES.md
@@ -20,7 +20,7 @@ For additional background, instructors (and students) can review the following n
 - [Collection of Jupyter Notebooks](https://github.com/jupyter/jupyter/wiki)
 - [The homogenization of scientific computing, or why Python is steadily eating other languages’ lunch](https://www.talyarkoni.org/blog/2013/11/18/the-homogenization-of-scientific-computing-or-why-python-is-steadily-eating-other-languages-lunch/)
 - [Shawn Rhoads - A brief introduction to Python for psychological science research](https://www.apa.org/science/about/psa/2019/07/python-research)
-- [R or Python for Psychologists](https://dominiquemakowski.github.io/post/r_or_python/)
+- [R or Python for Psychologists](https://dominiquemakowski.github.io/post/2020-05-22-r_or_python/)
 - [Russell Poldrack, et al. (2019) - Computational and informatics advances for reproducible data analysis in neuroimaging](https://www.annualreviews.org/doi/abs/10.1146/annurev-biodatasci-072018-021237)
 - [DartBrains - Intro to Programming](https://dartbrains.org/content/Introduction_to_Programming.html)
 - [DartBrains - Intro to Pandas](https://dartbrains.org/content/Introduction_to_Pandas.html)
