### 2022-01-11 2ec8ede
revised `paper.md`
Co-Authored-By: Lin Gan <63081264+gllg4009@users.noreply.github.com>
