+++ paper.md
@@ -26,13 +26,15 @@ authors:
 affiliations:
  - name: Georgetown University
    index: 1
-date: 27 September 2021
+date: 10 January 2022
 bibliography: paper.bib
 ---
 
 # Summary
 
-We present a semester-long educational course (see 14-week [schedule](https://shawnrhoads.github.io/gu-psyc-347/module-00-01_Course-Schedule.html)) and [Jupyter Book](https://shawnrhoads.github.io/gu-psyc-347) that provides introductory training in specifying, implementing, and interpreting computational models that characterize human social behavior and neuroscience. Through readings, discussions, and labs (e.g., Jupyter Notebook tutorials using Python), students will receive hands-on training in using mathematical models to test specific theories and hypotheses and explain unobservable aspects of complex social cognitive processes and behaviors. These aspects broadly include learning from and for others, learning about others, and social influences on decision-making and mental states. Background in introductory psychological science and basic research methods/statistics is highly recommended prior to participating in this course. Programming experience is helpful, but not required or expected. Although the course content is tailored for undergraduate students, it was developed to be easily adapted for trainees at higher career stages.
+We present a semester-long educational course (see 14-week [schedule](https://shawnrhoads.github.io/gu-psyc-347/module-00-01_Course-Schedule.html)) and [Jupyter Book](https://shawnrhoads.github.io/gu-psyc-347) that provides introductory training in specifying, implementing, and interpreting computational models that characterize human social behavior and neuroscience. Through readings, discussions, and labs (e.g., Jupyter Notebook tutorials using Python), students will receive hands-on training in using mathematical models to test specific theories and hypotheses and explain unobservable aspects of complex social cognitive processes and behaviors. These aspects broadly include learning from and for others, learning about others, and social influences on decision-making and mental states.
+
+Background in introductory psychological science and basic research methods/statistics is highly recommended prior to participating in this course. Programming experience is helpful, but not required or expected. Although the course content is tailored for undergraduate students, it was developed to be easily adapted for trainees at higher career stages.
 
 # Statement of Need
 
@@ -55,13 +57,15 @@ Throughout this course, students are expected to meet the following goals:
 
 The course begins with readings and discussions on the utility of computational modeling in psychological science and neuroscience (see [Statement of Need](#Statement-of-Need)) and training in scientific Python (**Module 01**). Students will learn how to set up an [Anaconda](https://anaconda.org/) Environment and work in [Google Colaboratory](https://colab.research.google.com) to run Python code in Executable Notebooks. They will then learn the basics of the Python programming language, including using `numpy` for scientific computing [@Numpy], `pandas` for data management [@Pandas], and `matplotlib` for data visualization [@Matplotlib].
 
-Then, students will learn the basics of modeling (**Module 02**). Using simulated data, they will learn how to fit linear and nonlinear models using least-squares minimization with a grid search, the `scipy.optimize.minimize` module [@Scipy], and the analytic solution for least-squares optimization. Using real data [@OConnell2021], they will learn how to estimate parameters (e.g., intercepts, slopes) from a linear model using both Ordinary Least Squares (OLS) and Maximum Likelihood Estimation (MLE). They will also learn how to assess model performance using $R^2$. They will compare their produced results to the actual results from the paper.
[…]
