### 2022-01-12 f2bc416
minor revisions to paper
