+++ paper.md
@@ -65,7 +65,7 @@ During the second half of the course, students will apply what they have learned
 
 All course materials were designed to be modular. For example, if students are knowledgeable in Python basics and Jupyter Notebook (**Module 01**), then teachers could begin with model fitting (**Module 02**). If students are knowledgeable in model fitting with maximum likelihood estimation (**Module 02**), teachers could start with reinforcement learning (**Module 03**). Any tutorial can be adopted, transformed, or built upon for other educational purposes (e.g., courses, single class sessions, workshops) under a [Creative Commons Attribution-ShareAlike 4.0 International License](https://github.com/shawnrhoads/gu-psyc-347/blob/master/LICENSE). Information on how teachers can use these materials can be found in the [README documentation](https://github.com/shawnrhoads/gu-psyc-347/blob/master/README.md).
 
-During the first iteration of the course in the spring 2021, students were 54.5% seniors (4th year), 36.4% juniors (3rd year), and 9.1% sophomores (2nd year). Broadly, students reported that they enrolled in the course to learn how to use Python and computational modeling to accelerate psychological research, to learn how to work with data and research methodology beyond a traditional research methods & statistics course, to learn skills relevant to a career in data science, and to gain a different perspective on topics from other classes. Most students reported spending 3-5 hours per weeks on readings and assignments (one student reported spending up to 10 hours and one student reported spending less than three hours). On a scale from 1 (strongly disagree) to 5 (strongly agree), students reported that the course was academically challenging and rigorous (*M*=4.75, *SD*=0.43) and that the course materials (*M*=4.00, *SD*=1.22), course-specific technologies (*M*=4.00, *SD*=1.22), and assignments (*M*=4.25, *SD*=0.83) helped them meet the stated learning objectives of the course.
+During the first iteration of the course in the spring 2021, students were 54.5% seniors (4th year), 36.4% juniors (3rd year), and 9.1% sophomores (2nd year). Broadly, students reported that they enrolled in the course to learn how to use Python and computational modeling to accelerate psychological research, to learn how to work with data and research methodology beyond a traditional research methods & statistics course, to learn skills relevant to a career in data science, and to gain a different perspective on topics from other classes. Most students reported spending 3-5 hours per week on readings and assignments (one student reported spending up to 10 hours and one student reported spending less than three hours). On a scale from 1 (strongly disagree) to 5 (strongly agree), students reported that the course was academically challenging and rigorous (*M*=4.75, *SD*=0.43) and that the course materials (*M*=4.00, *SD*=1.22), course-specific technologies (*M*=4.00, *SD*=1.22), and assignments (*M*=4.25, *SD*=0.83) helped them meet the stated learning objectives of the course.
 
 # Tutorials & Exercises
 
@@ -101,11 +101,11 @@ During the first iteration of the course in the spring 2021, students were 54.5%
 | Tutorial: Prosocial Learning | [Open in Colab](https://colab.research.google.com/github/shawnrhoads/gu-psyc-347/blob/master/docs/module-04-00_Social-Learning.ipynb) | [Open in Book](https://shawnrhoads.github.io/gu-psyc-347/module-04-00_Social-Learning.html) |
 | Exercise | [Open in Colab](https://colab.research.google.com/github/shawnrhoads/gu-psyc-347/blob/master/docs/module-04-01_Prosocial-RL-Exercises.ipynb) | [Open in Book](https://shawnrhoads.github.io/gu-psyc-347/module-04-01_Prosocial-RL-Exercises.html) |
 
[…]
