### 2022-01-12 29b4348
Delete blank_issue.md
