+++ .github/ISSUE_TEMPLATE/blank_issue.md
@@ -1,7 +1,0 @@
----
-name: Blank issue
-about: Create an issue without a template.
-title: ''
-labels: ''
-assignees: 'shawnrhoads'
----
