+++ paper.md
@@ -59,7 +59,7 @@ The course begins with readings and discussions on the utility of computational 
 
 Then, students will learn the basics of modeling (**Module 02**). Using simulated data, they will learn how to fit linear and nonlinear models using least-squares minimization with a grid search, the `scipy.optimize.minimize` module [@Scipy], and the analytic solution for least-squares optimization. Using real data [@OConnell2021], they will learn how to estimate parameters (e.g., intercepts, slopes) from a linear model using both Ordinary Least Squares (OLS) and Maximum Likelihood Estimation (MLE). They will also learn how to assess model performance using $R^2$ and the Bayesian Information Criterion. They will compare their produced results to the actual results from the paper.
 
-Next, students will learn the basics of reinforcement learning in non-social contexts (**Module 03**). They will learn how to run a "Two-Armed Bandit" experiment using PsychoPy [@Peirce2019] and explore behavioral data from the task. Following exercises adapted from [@Wilson2019] and implemented in Python, students will simulate learing using three models: (1) random responding, (2) noisy win-stay-lose-shift, and (3) Rescorla-Wagner reinforcement learning. They will use MLE to fit reinforcement learning models to simulated data and recover model parameters (e.g., learning rate, temperature) [@Palminteri2017].
+Next, students will learn the basics of reinforcement learning in non-social contexts (**Module 03**). They will learn how to run a "Two-Armed Bandit" experiment using PsychoPy [@Peirce2019] and explore behavioral data from the task. Following exercises adapted from [@Wilson2019] and implemented in Python, students will simulate learning using three models: (1) random responding, (2) noisy win-stay-lose-shift, and (3) Rescorla-Wagner reinforcement learning. They will use MLE to fit reinforcement learning models to simulated data and recover model parameters (e.g., learning rate, temperature) [@Palminteri2017].
 
 During the second half of the course, students will apply what they have learned in the context of social experiences and interactions [@Lockwood2020; @Zhang2019] with an increased focus on reading empirical research papers with computational models, presenting findings and interpreting results in journal club settings, and discussing with other students. Topics include learning about threats from others, learning how actions affect the outcomes of others, theory of mind, updating beliefs about others' states and traits, and social influences on decision-making. They will complete an adapted version of a social reinforcement learning task [@Lockwood2016] in PsychoPy [@Peirce2019] and develop models in Python to evaluate and compare the likelihood of each model given real data (**Module 04**). Finally, they will develop their own project and write a research proposal (formatted using guidelines for the National Science Foundation [Graduate Research Fellowship Program Award](https://nsfgrfp.org/) application).
 
@@ -103,13 +103,13 @@ During the first iteration of the course in the spring 2021, students were 54.5%
 
 Students have the option to interact with each tutorial notebook using one of two ways. The first option entails [installing Anaconda 3](https://docs.anaconda.com/anaconda/install/) on their local computers with a [custom environment](https://raw.githubusercontent.com/shawnrhoads/gu-psyc-347/master/course-env.yml) made for this course and then opening each notebook using Jupyter Notebook. Teachers can walk students through the installation steps using the instructions provided on the [Getting Started](https://shawnrhoads.github.io/gu-psyc-347/module-00-04_Getting-Started.html) page. However, this can be rather difficult to implement with students' different machines, operating systems, background, etc. As a fail-safe, students also have the option to use [Google Colaboratory](https://colab.research.google.com/), which runs the code on Google's cloud servers for **free**. There is a button at the top of each tutorial page that opens each Jupyter Book page in Google Colaboratory.
 
[…]
