### 2022-01-12 #146 @shawnrhoads
Hi all,

Thank you again for this helpful feedback! We have updated the paper and repository to reflect the following changes. For ease of review, we have separated our updates with two pull requests: [#184](https://github.com/shawnrhoads/gu-psyc-347/pull/184) updates the repository contents and [#186](https://github.com/shawnrhoads/gu-psyc-347/pull/186) updates the paper contents. Reviewers can view all changes directly using the hyperlinks above. We also summarize our changes below.

## Improving documentation for students, teachers, and contributors

With [#184](https://github.com/shawnrhoads/gu-psyc-347/pull/184), we have now separated the documentation for students, teachers, and contributors. These changes should also resolve [#180](https://github.com/shawnrhoads/gu-psyc-347/issues/180).

The repository documentation (`README.md`) now contains information for mainly for teachers/instructors and contributors. The Jupyter Book homepage (`docs/index.md`) now contains information targeted at students.

We also added a new resource page to the repository documentation (`RESOURCES.md`) to provide additional background for interested teachers/instructors (and students).

In addition, the `CONTRIBUTING.md` documentation now includes additional information on: (1) submitting issues, (2) requesting new content, (3) adding new content, (4) maintaining continuous integration, and (5) building the Jupyter Book locally.

To improve accessibility for students, teachers, and contributors, these updates also include the addition of [three new issue template forms](https://github.com/shawnrhoads/gu-psyc-347/issues/new/choose):
1. [Issue Reports](https://github.com/shawnrhoads/gu-psyc-347/issues/new?labels=bug&template=issue-template.yml)
2. [Enhancement Requests](https://github.com/shawnrhoads/gu-psyc-347/issues/new?labels=enhancement&template=enhancement-template.yml)
3. [Content-Relevant Questions](https://github.com/shawnrhoads/gu-psyc-347/issues/new?labels=question&template=question-template.yml)

With [#186](https://github.com/shawnrhoads/gu-psyc-347/pull/186), we updated the paper to reflect the above changes with a focus on how teachers may adopt the course modules:

**Learning Objectives**
> All course materials were designed to be modular. For example, if students are knowledgeable in Python basics and Jupyter Notebook (**Module 01**), then teachers could begin with model fitting (**Module 02**). If students are knowledgeable in model fitting with maximum likelihood estimation (**Module 02**), teachers could start with reinforcement learning (**Module 03**). Any tutorial can be adopted, transformed, or built upon for other educational purposes (e.g., courses, single class sessions, workshops) under a [Creative Commons Attribution-ShareAlike 4.0 International License](https://github.com/shawnrhoads/gu-psyc-347/blob/master/LICENSE). Information on how teachers can use these materials can be found in the [README documentation](https://github.com/shawnrhoads/gu-psyc-347/blob/master/README.md).

**Tutorials & Exercises**
> Students have the option to interact with each tutorial notebook using one of two ways. The first option entails [installing Anaconda 3](https://docs.anaconda.com/anaconda/install/) on their local computers with a [custom environment](https://raw.githubusercontent.com/shawnrhoads/gu-psyc-347/master/course-env.yml) made for this course and then opening each notebook using Jupyter Notebook. Teachers can walk students through the installation steps using the instructions provided on the [Getting Started](https://shawnrhoads.github.io/gu-psyc-347/module-00-04_Getting-Started.html) page. However, this can be rather difficult to implement with students' different machines, operating systems, background, etc. As a fail-safe, students also have the option to use [Google Colaboratory](https://colab.research.google.com/), which runs the code on Google's cloud servers for **free**. There is a button at the top of each tutorial page that opens each Jupyter Book page in Google Colaboratory.
>
> Once students are able to run notebooks (locally or on the cloud), they can interact with all of the code in the tutorials. Teachers should guide students through all of the "tutorials", which, for example, will allow them to demonstrate to students how tweaking certain variables in real-time would change the outputs. The "exercises" at the end of each module were created for the students to apply what they have learned. They can work on these by themselves or in groups. Links to example solutions for all exercises are provided on each page. In addition, teachers should familiarize themselves with the [recommended readings](https://shawnrhoads.github.io/gu-psyc-347/module-00-03_Reading-List.html) associated with each module and use them to introduce students to each topic, generate discussions and ideas among students, and provide relevant background for each tutorial. For example, the Wilson & Collins (2019) paper can be assigned for reading and discussion prior to beginning any notebook tutorials for **Module 03**.
>
> For additional background on the content covered in each tutorial (technical or theoretical), teachers and students are encouraged to review the references and/or resources listed at the beginning of each tutorial or the non-exhaustive list of resources on the [Resources page](https://github.com/shawnrhoads/gu-psyc-347/blob/master/RESOURCES.md). Content-related questions from students, teachers, and contributors related to the course can also be submitted directly on the GitHub repository. However, we request that individuals make a good faith effort searching for their answer in the relevant readings or resources listed at the beginning of each tutorial before submitting questions.
>

**Future Directions**
> We appreciate any feedback on errors, omitted credit, or new ideas for this evolving project. These can be communicated on GitHub by either [reporting an issue](https://github.com/shawnrhoads/gu-psyc-347/issues/new?labels=bug&template=issue-template.yml) or [requesting an enhancement](https://github.com/shawnrhoads/gu-psyc-347/issues/new?labels=enhancement&template=enhancement-template.yml). All contributions will be credited and should follow the community guidelines outlined on the [Contribution page](https://shawnrhoads.github.io/gu-psyc-347/module-00-06_Contributing.html).

<hr>

## Experience/Story and Expertise

With [#186](https://github.com/shawnrhoads/gu-psyc-347/pull/186), we have now updated the following sections in the paper regarding students' expectations and experiences related to the learning objectives as well as the relevant expertise/story of the authors.

**Learning Objectives**
> During the first iteration of the course in the spring 2021, students were 54.5% seniors (4th year), 36.4% juniors (3rd year), and 9.1% sophomores (2nd year). Broadly, students reported that they enrolled in the course to learn how to use Python and computational modeling to accelerate psychological research, to learn how to work with data and research methodology beyond a traditional research methods & statistics course, to learn skills relevant to a career in data science, and to gain a different perspective on topics from other classes. Most students reported spending 3-5 hours per week on readings and assignments (one student reported spending up to 10 hours and one student reported spending less than three hours). On a scale from 1 (strongly disagree) to 5 (strongly agree), students reported that the course was academically challenging and rigorous (*M*=4.75, *SD*=0.43) and that the course materials (*M*=4.00, *SD*=1.22), course-specific technologies (*M*=4.00, *SD*=1.22), and assignments (*M*=4.25, *SD*=0.83) helped them meet the stated learning objectives of the course.

**Author Contributions**
> [S. A. R.](https://github.com/shawnrhoads) learned to implement the relevant course methods through conducting research and participating in various summer courses (e.g., [2018 MIND Computational Summer School](https://mindsummerschool.org/), 2018 Neuroscience School of Advanced Studies in Social Decision-Making, [2019 Computational Psychiatry Summer Course](https://www.cpcourse.org/)). He also received relevant pedagogical training as a Teaching Assistant (TA) for [Neuromatch Academy](https://academy.neuromatch.io/), as a TA for Dr. Maximilian Risenhuber's Computational Neuroscience course, and as a participant in the [Apprenticeship in Teaching Program](https://cndls.georgetown.edu/atprogram/) at Georgetown University. The course materials were created to make the hands-on elements from these experiences as well as smaller, related workshops led by [S. A. R.](https://github.com/shawnrhoads) more accessible in university settings and online.

Thanks again and please reach out if you have questions or additional comments.

Best,
Shawn & Lin
