### 2022-01-16 #146 @djmannion
I'm happy to approve the revised paper. Thanks to the authors for creating this resource and writing it up!

A few other general comments below.

> Usage: Does the documentation explain how someone would adopt the module, and include examples of how to use it?

The documentation on how to build the course site is clear and I was able to build it successfully (though it is missing the Contributing document, which seems to require the build to go via github).

I noticed that there are quite a few locations where the build process is specific to the course. For example, the `_config.yml` has things like the course URL, a Google analytics ID, the repository URL, etc. - and a lot of Module 0 is specific to this particular course. Given that it is stated in the documentation that: "Instructors can adopt, remix, transform, and build upon these materials for their own course or for other educational purpose", I was going to suggest that the authors add a statement about this specificity and perhaps some instructions on what would need changing for other instructors to be able to convert it to their own course. However, I think it is clear enough from the context that what is presented is for a specific implementation of the material and that some digging would be needed to adapt it to use in a different course.

## Miscellaneous

These are just some general thoughts I had when (lightly) reading through the course material. They are outside the scope of the review context, so it isn't necessary to respond to these or make any changes in response - I just thought I'd pass them along.

* A couple of typos in the Modules section of "Intro to Python Basics": "python import module_name", "python import module_name as new_module_name".
* I wasn't sure why some notebooks show the cell outputs (e.g., "Working with Data", "Linear Modeling") while others do not (e.g., "Intro to Python Basics", "Nonlinear Modeling) - in the built site, that is.
* You might not want to add in any extra syntax confusion, but the extraneous cell output that gets produced sometimes by matplotlib (e.g. `<matplotlib.legend.Legend at 0x174e572d908>`) can be suppressed by adding a semi-colon to the end of the command (e.g., `plt.legend();`) - in case you weren't aware of that.
* I like the use of `NotImplementedError` in the exercises!
* It might be worth applying a line-length restriction to the code - it is a bit awkward to horizontally scroll.
* The colour scale for the heatmap of the "Least-squares minimization" section makes it difficult to see the minimum (on my screen, anyway) - there are many cells that are black.
* Regarding the "Pedagogy: Does the module seem easy to follow? Does it observe guidance on cognitive load?" criterion, it did seem to me that there are a *lot* of new concepts introduced in Module 1. I realise that achieving the right balance between covering foundational concepts and applied examples is difficult, though (I have struggled with it in my teaching, at least). Anyway, maybe you could consider in the future whether all that material is necessary - for example, I didn't see much usage of custom classes outside of Module 1 (though I may have missed it).
