### 2022-01-24 #146 @emckiernan
Super, thanks @djmannion and @arokem for your reviews! We're very grateful.

@shawnrhoads: Looks like we're almost there! There's just a few minor comments above from @djmannion that I'd suggest looking over. As mentioned, it's not necessary to address all of these, but I think some would definitely be worth thinking about/doing. Just let me know when you've got the final version, and then I think we'll be good to go on final acceptance!
