+++ docs/_config.yml
@@ -16,8 +16,8 @@ only_build_toc_files: false
 # Execution settings
 execute:
   timeout: 5000
-  execute_notebooks: cache  # Whether to execute notebooks at build time. Must be one of ("auto", "force", "cache", "off")
-  cache: "docs/.jupyter_cache/"  # A path to the jupyter cache that will be used to store execution artifacs. Defaults to `_build/.jupyter_cache/`
+  execute_notebooks: auto  # Whether to execute notebooks at build time. Must be one of ("auto", "force", "cache", "off")
+  cache: ""  # A path to the jupyter cache that will be used to store execution artifacs. Defaults to `_build/.jupyter_cache/` # "docs/.jupyter_cache/"
   exclude_patterns: []  # A list of patterns to *skip* in execution (e.g. a notebook that takes a really long time)
   stderr_output: remove
 #######################################################################################
