### 2022-01-24 #146 @shawnrhoads
Sounds great, thank you all so much for these reviews!

I just merged a few updates with [#187](https://github.com/shawnrhoads/gu-psyc-347/pull/187) addressing the following minor suggestions made by @djmannion:

> A couple of typos in the Modules section of "Intro to Python Basics": "python import module_name", "python import module_name as new_module_name".

- Thanks for catching these! We now fixed these (see [here](https://shawnrhoads.github.io/gu-psyc-347/module-01-01_Intro-to-Python.html#modules)).

> I wasn't sure why some notebooks show the cell outputs (e.g., "Working with Data", "Linear Modeling") while others do not (e.g., "Intro to Python Basics", "Nonlinear Modeling) - in the built site, that is.

- Thanks for catching this, too. This was the result from a more recent change with the configuration settings. These pages now include outputs: [here](https://shawnrhoads.github.io/gu-psyc-347/module-01-01_Intro-to-Python.html) and [here](https://shawnrhoads.github.io/gu-psyc-347/module-02-01_Nonlinear-Modeling.html).

> You might not want to add in any extra syntax confusion, but the extraneous cell output that gets produced sometimes by matplotlib (e.g. <matplotlib.legend.Legend at 0x174e572d908>) can be suppressed by adding a semi-colon to the end of the command (e.g., plt.legend();) - in case you weren't aware of that.

- Great idea! We added this change [here](https://shawnrhoads.github.io/gu-psyc-347/module-01-02_Working-with-Data.html#visualizing-data-using-matplotlib).

> The colour scale for the heatmap of the "Least-squares minimization" section makes it difficult to see the minimum (on my screen, anyway) - there are many cells that are black.

- We changed the colormap and also now added a rectangle around the minimum (see [here](https://shawnrhoads.github.io/gu-psyc-347/module-02-00_Linear-Modeling.html#least-squares-minimization)).
