### 2022-01-24 17e5f18
updated emoji spacing
