+++ README.md
@@ -2,11 +2,11 @@
 
 # Computational Models of Human Social Behavior and Neuroscience
 
-## For Students 🎓
+## 🎓 For Students
 
 All course information relevant for students can be found on the <a target="_blank" rel="noopener noreferrer" href="https://shawnrhoads.github.io/gu-psyc-347">course homepage</a> and <a target="_blank" rel="noopener noreferrer" href="https://shawnrhoads.github.io/gu-psyc-347/module-00-00_Syllabus.html">syllabus</a>. Students are free to follow along on their own or with an instructor (see below).
 
-## For Instructors 🍎
+## 🍎 For Instructors
 
 This repository contains materials for a semester-long educational course (see 14-week <a target="_blank" rel="noopener noreferrer" href="https://shawnrhoads.github.io/gu-psyc-347/module-00-01_Course-Schedule.html">schedule</a>) and <a target="_blank" rel="noopener noreferrer" href="https://shawnrhoads.github.io/gu-psyc-347">Jupyter Book</a> that provides introductory training in specifying, implementing, and interpreting computational models that characterize human social behavior and neuroscience. This course was designed and taught by <a target="_blank" rel="noopener noreferrer" href="https://shawnrhoads.github.io">Shawn A Rhoads</a> in the spring of 2021 (reach him via <a target="_blank" rel="noopener noreferrer" href="mailto:sr1209(at)georgetown(dot)edu">email</a>, <a target="_blank" rel="noopener noreferrer" href="https://github.com/shawnrhoads">GitHub</a>, or <a target="_blank" rel="noopener noreferrer" href="https://twitter.com/intent/user?user_id=380634496">Twitter</a> with any questions).
 
@@ -56,7 +56,7 @@ Once students are able to run notebooks (locally or on the cloud), they can inte
 
 For additional background on the content covered in each tutorial (technical or theoretical), instructors (and students) are encouraged to review the references and/or resources listed at the beginning of each tutorial or the non-exhaustive list of resources on the <a target="_blank" rel="noopener noreferrer" href="https://github.com/shawnrhoads/gu-psyc-347/blob/master/RESOURCES.md">Resources page</a>. In addition, instructors should familarize themselves with the <a target="_blank" rel="noopener noreferrer" href="https://shawnrhoads.github.io/gu-psyc-347/module-00-03_Reading-List.html">recommended readings</a> associated with each module and use them to introduce students to each topic, generate discussions and ideas among students, and provide relevant background for each tutorial.
 
-## For Contributors 🤝
+## 🤝 For Contributors
 
 While the course aims to incorporate a diverse curriculum to introduce students to the basics of modeling in the context of social behavior, there is still so much left out (and still much unknown in this growing field). That being said, other researchers and educators are invited to help improve and expand the content included in this Jupyter Book!
 
@@ -67,13 +67,13 @@ Here are some ways you can help accomplish this goal!
 
 Please visit <a target="_blank" rel="noopener noreferrer" href="https://shawnrhoads.github.io/gu-psyc-347/module-00-06_Contributing.html">this page</a> for additional information on helping improve and/or expand the content in this Jupyter Book!
 
-## Seeking Help 🔎
+## 🔎 Seeking Help
 
[…]
