+++ _sources/index.md
@@ -3,7 +3,7 @@
 # Computational Models of Human Social Behavior and Neuroscience
 The content in this Jupyter Book is subject to change. 
 
-## Course information 📝
+## 📝 Course information
 **Section**: PSYC 347-01</br>
 **Prerequisites**: PSYC 002 - Research Methods and Statistics (or equivalent); No prior programming experience necessary</br>
 **Required materials**: Only a working laptop/computer is needed! Book chapters, journal articles, and software are openly available to all students (no purchase required)</br>
@@ -13,7 +13,7 @@ The content in this Jupyter Book is subject to change.
 **Syllabus**: <a target="_blank" rel="noopener noreferrer" href="https://shawnrhoads.github.io/gu-psyc-347/module-00-00_Syllabus.html">Found here</a></br>
 **Schedule**: <a target="_blank" rel="noopener noreferrer" href="https://shawnrhoads.github.io/gu-psyc-347/module-00-01_Course-Schedule.html">Found here</a>
 
-## Instructor information 👨‍🏫
+## 👨‍🏫 Instructor information
 **Instructor**: Shawn A Rhoads<br>
 **Pronouns**: he/him/his<br>
 **Email**: sr1209(at)georgetown(dot)edu<br>
@@ -22,7 +22,7 @@ The content in this Jupyter Book is subject to change.
 
 Other instructors are encouraged to adopt, remix, transform, and/or build upon the following materials for educational purposes under a <a rel="license" target="_blank" rel="noopener noreferrer" href="http://creativecommons.org/licenses/by-sa/4.0/">Creative Commons Attribution-ShareAlike 4.0 International License</a>. Information for instructors can be found in the <a target="_blank" rel="noopener noreferrer" href="https://github.com/shawnrhoads/gu-psyc-347">README documentation</a>.
 
-## Tutorials & Exercises 🏋️
+## 🏋️ Tutorials & Exercises
 
 ### Module 01 - Scientific Python
 
@@ -56,13 +56,13 @@ Other instructors are encouraged to adopt, remix, transform, and/or build upon t
 | Tutorial: Prosocial Learning | <a target="_blank" rel="noopener noreferrer" href="https://colab.research.google.com/github/shawnrhoads/gu-psyc-347/blob/master/docs/module-04-00_Social-Learning.ipynb">![Open In Colab](https://colab.research.google.com/assets/colab-badge.svg)</a> | <a target="_blank" rel="noopener noreferrer" href="https://shawnrhoads.github.io/gu-psyc-347/module-04-00_Social-Learning.html">![view](https://jupyterbook.org/badge.svg)</a> |
 | Exercise | <a target="_blank" rel="noopener noreferrer" href="https://colab.research.google.com/github/shawnrhoads/gu-psyc-347/blob/master/docs/module-04-01_Prosocial-RL-Exercises.ipynb">![Open In Colab](https://colab.research.google.com/assets/colab-badge.svg)</a> | <a target="_blank" rel="noopener noreferrer" href="https://shawnrhoads.github.io/gu-psyc-347/module-04-01_Prosocial-RL-Exercises.html">![view](https://jupyterbook.org/badge.svg)</a> |
 
-## Acknowledgments 🙏
+## 🙏 Acknowledgments
[…]
