### 2022-01-26 #146 @emckiernan
@shawnrhoads: Super, thanks, we are good for final acceptance! Congrats!

This is the first time I get to this stage as a JOSE editor, so please bear with me :) In the guidelines it says, "When a submission is accepted, we ask that the authors create an archive (on Zenodo, figshare, or other) and post the archive DOI in the REVIEW issue". Could you please do this and post the DOI here? I can then reach out to the EiC for final processing of your article. Thanks in advance!
