### 2022-01-26 #146 @labarba
First make a tagged release in the repository, and post the version number here.

Erin, you will then run `@whedon set v1.2.3 as version` here (with the correct version number).
