### 2022-01-26 #146 @shawnrhoads
Incredible! Thanks so much!

The latest tagged release is now `v1.1.0`
