### 2022-01-26 #146 @shawnrhoads
On Zenedo, there is an option that says the following:

> Optional. Did your publisher already assign a DOI to your upload? If not, leave the field empty and we will register a new DOI for you. A DOI allows others to easily and unambiguously cite your upload. Please note that it is NOT possible to edit a Zenodo DOI once it has been registered by us, while it is always possible to edit a custom DOI.

Should I provide a DOI from JOSE or have one assigned by Zenodo?
