### 2022-01-26 #146 @labarba
Have Zenodo assign a DOI. This is for the archive of your materials. The JOSE DOI is for the citable paper only.
