### 2022-01-26 #146 @emckiernan
@labarba Hi Lorena, thanks for the assist :) I think everything is good to go here! I've set the version and archive DOI, checked the proofs and refs, etc. Is there anything else for me to do?
