### 2022-01-26 #146 @labarba
👋 If everything is ready to go, you as handling editor run `@whedon recommend-accept` — and I think that would automatically do the check on references and generate the pdf ... try it! 😁
