### 2022-01-26 #146 @whedon
:wave: @openjournals/jose-eics, this paper is ready to be accepted and published.
